Introduced GradientFeatures. Updated ExpressionMatrix, MatrixFunctionEvaluator and SequenceFunctionEvaluator to use a lower value for the sequence limit. Updated Expression::init so that nodeCount is not allowed to exceed the size of an Integer. Improvements to alert features in Console and introduced a maximum for total messages sent via SignalWireDeliveryProvider. Test updates.

+++ code/src/main/java/io/almostrealism/expression/Expression.java
@@ -105,7 +105,7 @@ public abstract class Expression<T> implements
 		ScopeSettings.reviewChildren(getChildren());
 
 		this.depth = getChildren().stream().mapToInt(e -> e.depth).max().orElse(-1) + 1;
-		this.nodeCount = getChildren().stream().mapToInt(e -> e.nodeCount).sum() + 1;
+		this.nodeCount = Math.toIntExact(getChildren().stream().mapToLong(e -> e.nodeCount).sum() + 1);
 		this.containsLong = (getType() == Long.class ||
 				getChildren().stream().anyMatch(e -> e.containsLong))
 				&& intValue().isEmpty();
